maps redone for stretch and squash
final submission for a2

+++ graphics.c
@@ -64,8 +64,8 @@ GLfloat viewpointLight[] = {-50.0, -50.0, -50.0, 1.0};
 float skySize;
 
 	/* screen dimensions */
-int screenWidth = 1024;
-int screenHeight = 768;
+float screenWidth = 1024;
+float screenHeight = 768;
 
 	/* command line flags */
 int flycontrol = 1;		// allow viewpoint to move in y axis when 1
